# backend/openapi.yaml
...
openapi: 3.1.0
info:
  title: Shift Scheduler & Attendance API
  version: 2.4.0
  description: "Backend of docs/SPEC.md v2.4. Error envelope: { error: { code, message, details?, requestId } }. Lists: { data, meta: { page, limit, total } }. Calendar endpoints are range based (max 62 days)."
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    deviceToken:
      type: apiKey
      in: header
      name: X-Device-Token
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - requestId
          properties:
            code:
              type: string
              example: EMPLOYEE_ON_TIME_OFF
            message:
              type: string
            details:
              type: array
              items:
                type: object
            requestId:
              type: string
...
    Shift:
      type: object
      properties:
        id:
          type: integer
        hotelId:
          type: integer
        departmentId:
          type: integer
        name:
          type: string
        startTime:
          type: string
          example: 06:00
        endTime:
          type: string
        durationHours:
          type: number
        breakDurationMinutes:
          type: integer
        paidHours:
          type: number
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/Warning"
    Warning:
      type: object
      properties:
        type:
          type: string
        severity:
          enum:
            - warning
            - info
        message:
          type: string
      additionalProperties: true
...
paths:
...
  /departments:
    get:
      operationId: D1
      tags:
        - Structure
      summary: "D1: List departments"
      description: "Role: staff and up."
      security:
        - bearer: []
      parameters:
        - name: hotelId
          in: query
          required: false
          schema:
            type: string
          description: Hotel id; required when the caller can access more than one hotel
        - name: page
          in: query
          required: false
          schema:
            type: string
          description: Page (default 1)
        - name: limit
          in: query
          required: false
          schema:
            type: string
          description: Page size (default 50, max 100)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: UNAUTHENTICATED / DEVICE_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: RESOURCE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: D2
      tags:
        - Structure
      summary: "D2: Create department"
      description: "Role: manager and up."
      security:
        - bearer: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                type: object
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: UNAUTHENTICATED / DEVICE_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: RESOURCE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /departments/{id}:
    patch:
      operationId: D3
      tags:
        - Structure
      summary: "D3: Update department"
      description: "Role: manager and up."
      security:
        - bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Numeric id ('me' allowed for /employees/{id}/...)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: UNAUTHENTICATED / DEVICE_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: RESOURCE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "412":
          description: PRECONDITION_FAILED (If-Match)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: D4
      tags:
        - Structure
      summary: "D4: Soft-delete (409 while in use)"
      description: "Role: manager and up."
      security:
        - bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Numeric id ('me' allowed for /employees/{id}/...)
      responses:
        "204":
          description: No content
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: UNAUTHENTICATED / DEVICE_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: RESOURCE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /shifts:
    get:
      operationId: S1
      tags:
        - Structure
      summary: "S1: List shifts"
      description: "Role: staff and up."
      security:
        - bearer: []
      parameters:
        - name: hotelId
          in: query
          required: false
          schema:
            type: string
          description: Hotel id; required when the caller can access more than one hotel
        - name: departmentId
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: string
          description: Page (default 1)
        - name: limit
          in: query
          required: false
          schema:
            type: string
          description: Page size (default 50, max 100)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: UNAUTHENTICATED / DEVICE_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: RESOURCE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: S2
      tags:
        - Structure
      summary: "S2: Create shift (break warnings)"
      description: "Role: admin."
      security:
        - bearer: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: UNAUTHENTICATED / DEVICE_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: RESOURCE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /shifts/{id}:
    patch:
      operationId: S3
      tags:
        - Structure
      summary: "S3: Update shift"
      description: "Role: admin."
      security:
        - bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Numeric id ('me' allowed for /employees/{id}/...)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: UNAUTHENTICATED / DEVICE_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: RESOURCE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "412":
          description: PRECONDITION_FAILED (If-Match)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: S4
      tags:
        - Structure
      summary: "S4: Soft-delete (409 with future entries)"
      description: "Role: admin."
      security:
        - bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Numeric id ('me' allowed for /employees/{id}/...)
      responses:
        "204":
          description: No content
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: UNAUTHENTICATED / DEVICE_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: RESOURCE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /shifts/{id}/staffing-requirements:
    get:
      operationId: S5
      tags:
        - Structure
      summary: "S5: Minimum staff per ISO weekday"
      description: "Role: manager and up."
      security:
        - bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Numeric id ('me' allowed for /employees/{id}/...)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: UNAUTHENTICATED / DEVICE_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: RESOURCE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: S6
      tags:
        - Structure
      summary: "S6: Replace staffing requirements"
      description: "Role: manager and up."
      security:
        - bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Numeric id ('me' allowed for /employees/{id}/...)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: UNAUTHENTICATED / DEVICE_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: RESOURCE_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...